+++ src/window.rs
@@ -384,6 +384,23 @@ impl Window {
         self.window.get_current_monitor()
     }
 
+    /// Returns the list of all the monitors available on the system.
+    ///
+    /// This is the same as `EventsLoop::get_available_monitors`, and is provided for convenience.
+    #[inline]
+    pub fn get_available_monitors(&self) -> AvailableMonitorsIter {
+        let data = self.window.get_available_monitors();
+        AvailableMonitorsIter { data: data.into_iter() }
+    }
+
+    /// Returns the primary monitor of the system.
+    ///
+    /// This is the same as `EventsLoop::get_primary_monitor`, and is provided for convenience.
+    #[inline]
+    pub fn get_primary_monitor(&self) -> MonitorId {
+        MonitorId { inner: self.window.get_primary_monitor() }
+    }
+
     #[inline]
     pub fn id(&self) -> WindowId {
         WindowId(self.window.id())
@@ -393,6 +410,7 @@ impl Window {
 /// An iterator for the list of available monitors.
 // Implementation note: we retrieve the list once, then serve each element by one by one.
 // This may change in the future.
+#[derive(Debug)]
 pub struct AvailableMonitorsIter {
     pub(crate) data: VecDequeIter<platform::MonitorId>,
 }